04.30
- Host auth
- Sub view in MainMenu
- Game state
- Join | create | exit game
- Live player & game update

+++ src/game/GameManager.ts
@@ -1,10 +1,12 @@
 import { initializeApp } from "firebase/app";
-import { getFirestore, CollectionReference, collection, getDocs, DocumentReference, doc, setDoc, deleteDoc, onSnapshot, DocumentSnapshot } from "firebase/firestore";
+import { getFirestore, CollectionReference, collection, getDocs, DocumentReference, doc, setDoc, deleteDoc, onSnapshot, getDoc } from "firebase/firestore";
 import { Unsubscribe, UserCredential, deleteUser, getAuth, signInAnonymously } from "firebase/auth";
 import { PlayerData } from "./types/PlayerData";
 import { GameData } from "./types/GameData";
 import { QuizList } from "./types/QuizData";
 
+
+//#region Initialize Firebase
 const firebaseConfig = {
     apiKey: "AIzaSyBiupN7TDtDoYptirlRlzH2AQpIPdjCIrI",
     authDomain: "quiz-game-a283c.firebaseapp.com",
@@ -14,67 +16,24 @@ const firebaseConfig = {
     messagingSenderId: "559679834708",
     appId: "1:559679834708:web:3d25c3e66148152d8abbc1",
     measurementId: "G-82X1GE6N5C"
-  };
+};
   
-//#region Initialize Firebase
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
 const auth = getAuth();
 const quizListRef: CollectionReference = collection(db, "quizList");
-const playersRef: CollectionReference = collection(db, "players");
-
-/*
-//#region sign in & sign out
-const signInIntoApp = async (name: string): Promise<Player | undefined> => {
-
-    const player: UserCredential = await signInAnonymously(auth);
-
-    if(player) {
-        const playerData: Player = {
-            name: name,
-            score: 0
-        };
-
-        playerRef = doc(collection(db, "players"));
-        await setDoc(playerRef, playerData);
-        
-        return playerData;
-    };
-
-    return undefined;
-};
-const signOutFromApp = async (): Promise<void> => {
-    await deleteDoc(playerRef);
-    if(auth.currentUser)
-        await deleteUser(auth.currentUser);
-};
 //#endregion
 
-//#region game action
-const createGame = async (info: QuizInfo): Promise<GameData | undefined> => {
-    if(!info) return undefined;
-
-};
-//#endregion
-
-//#region quiz action
-const quizListFetcher = async (): Promise<QuizList> => {
-    const docSnap: QuerySnapshot<DocumentData> = await getDocs(quizListRef);
-
-    if(!docSnap.empty)
-        return docSnap.docs.map(doc => doc.data()) as QuizList;
-
-    return [];
-};
-//#endregion
-*/
-
+//#region references & datas
 let localPlayer: PlayerData;
+let localPlayerIndex: number = -1;
 let game: GameData;
 let gameRef: DocumentReference;
 let gameUnSub: Unsubscribe;
 let quizList: QuizList;
+//#endregion
 
+//#region authentication
 const signIn = async (name: string): Promise<PlayerData | undefined> => {
     const player: UserCredential = await signInAnonymously(auth);
 
@@ -93,6 +52,7 @@ const signOut = async (): Promise<boolean> => {
     }
     return false;
 }
+//#endregion
 
 const getQuizList = async (): Promise<QuizList> => {
     const querySnapshot = await getDocs(quizListRef);
@@ -103,7 +63,40 @@ const getQuizList = async (): Promise<QuizList> => {
     return quizList;
 };
 
-const createGame = async (data: GameData, event: (data: GameData, id: string) => void): Promise<boolean> => {
+//#region Game connection functions
+const joinGame = async (id: string, event: (data: GameData, id: string) => void): Promise<GameData> => {
+    const gameDoc = await getDoc(doc(db, "games", id));
+
+    if(gameDoc.exists()){
+        gameRef = doc(db, "games", id);
+        game = gameDoc.data() as GameData;
+        
+        if(!game.players){
+            game = {
+                ...game,
+                players: [localPlayer]
+            };
+            localPlayerIndex = 0;
+        }else{
+            game.players.push(localPlayer);
+            localPlayerIndex = game.players.length - 2;
+        }
+        await setDoc(gameRef, game);
+        
+        gameUnSub = onSnapshot(gameRef, doc => {
+            if(doc.exists()){
+                game = doc.data();
+                event(game, id);
+            }
+        });
+
+        return game;
+    }else{
+        return {} as GameData;
+    }
+};
+
+const createGame = async (data: GameData, event: (data: GameData, id: string) => void): Promise<void> => {
     game = data;
     gameRef = doc(collection(db, "games"));
     await setDoc(gameRef, game);
@@ -112,21 +105,27 @@ const createGame = async (data: GameData, event: (data: GameData, id: string) =>
        if(doc.exists())
             event(doc.data(), doc.id);
     });
-
-    return true;
 };
 
-const exitGame = async (): Promise<void> => {
-    game = {};
+const exitGame = async (destroy?: boolean): Promise<void> => {
     gameUnSub();
-    await deleteDoc(gameRef);
+    if(destroy){
+        await deleteDoc(gameRef);
+    }else{
+        game.players = game.players?.splice(localPlayerIndex, 1);
+        await setDoc(gameRef, game);
+    }
+    localPlayerIndex = -1;
+    game = {};
     gameRef = {} as DocumentReference;
 };
+//#endregion
 
 export default {
     signIn,
     signOut,
     getQuizList,
     createGame,
-    exitGame
+    exitGame,
+    joinGame
 };
